+++ bungee/src/main/java/com/github/games647/fastlogin/bungee/MojangApiBungee.java
@@ -5,13 +5,16 @@ import com.github.games647.fastlogin.core.MojangApiConnector;
 
 import java.util.List;
 import java.util.UUID;
+import java.util.concurrent.ConcurrentMap;
+import java.util.logging.Logger;
 
 import net.md_5.bungee.BungeeCord;
 
 public class MojangApiBungee extends MojangApiConnector {
 
-    public MojangApiBungee(FastLoginCore plugin, List<String> localAddresses, boolean apiLookup) {
-        super(plugin, localAddresses, apiLookup);
+    public MojangApiBungee(ConcurrentMap<Object, Object> requests, Logger logger, List<String> localAddresses
+            , int rateLimit) {
+        super(requests, logger, localAddresses, rateLimit);
     }
 
     @Override
@@ -24,11 +27,5 @@ public class MojangApiBungee extends MojangApiConnector {
     public boolean hasJoinedServer(Object session, String serverId) {
         //this is not needed in Bungee
         throw new UnsupportedOperationException("Not supported");
-    }
-
-    @Override
-    protected UUID getUUIDFromJsonAPI(String json) {
-        McAPIProfile apiPlayer = BungeeCord.getInstance().gson.fromJson(json, McAPIProfile.class);
-        return FastLoginCore.parseId(apiPlayer.getUuid());
     }
 }