+++ core/src/main/java/com/github/games647/fastlogin/core/MojangApiConnector.java
@@ -11,7 +11,9 @@ import java.util.HashSet;
 import java.util.List;
 import java.util.Set;
 import java.util.UUID;
+import java.util.concurrent.ConcurrentMap;
 import java.util.logging.Level;
+import java.util.logging.Logger;
 import java.util.regex.Pattern;
 
 import javax.net.ssl.HttpsURLConnection;
@@ -22,7 +24,7 @@ public abstract class MojangApiConnector {
     private static final int TIMEOUT = 1 * 1_000;
     private static final String USER_AGENT = "Premium-Checker";
 
-    private static final String MCAPI_UUID_URL = "https://us.mc-api.net/v3/uuid/";
+    private static final String MCAPI_UUID_URL = "https://mcapi.ca/uuid/player/";
 
     //only premium (paid account) users have a uuid from here
     private static final String UUID_LINK = "https://api.mojang.com/users/profiles/minecraft/";
@@ -34,15 +36,23 @@ public abstract class MojangApiConnector {
     //compile the pattern only on plugin enable -> and this have to be threadsafe
     private final Pattern playernameMatcher = Pattern.compile(VALID_PLAYERNAME);
 
+    private final ConcurrentMap<Object, Object> requests;
     private final BalancedSSLFactory sslFactory;
-    private final boolean apiLookup;
+    private final int rateLimit;
     private long lastRateLimit;
 
-    protected final FastLoginCore plugin;
+    protected final Logger logger;
 
-    public MojangApiConnector(FastLoginCore plugin, List<String> localAddresses, boolean apiLookup) {
-        this.plugin = plugin;
-        this.apiLookup = apiLookup;
+    public MojangApiConnector(ConcurrentMap<Object, Object> requests, Logger logger, List<String> localAddresses
+            , int rateLimit) {
+        this.logger = logger;
+        this.requests = requests;
+        
+        if (rateLimit > 600) {
+            this.rateLimit = 600;
+        } else {
+            this.rateLimit = rateLimit;
+        }
 
         if (localAddresses.isEmpty()) {
             this.sslFactory = null;
@@ -52,13 +62,13 @@ public abstract class MojangApiConnector {
                 try {
                     InetAddress address = InetAddress.getByName(localAddress);
                     if (!address.isAnyLocalAddress()) {
-                        plugin.getLogger().log(Level.WARNING, "Submitted IP-Address is not local", address);
+                        logger.log(Level.WARNING, "Submitted IP-Address is not local", address);
                         continue;
                     }
 
                     addresses.add(address);
                 } catch (UnknownHostException ex) {
-                    plugin.getLogger().log(Level.SEVERE, "IP-Address is unknown to us", ex);
+                    logger.log(Level.SEVERE, "IP-Address is unknown to us", ex);
                 }
             }
 
@@ -76,10 +86,12 @@ public abstract class MojangApiConnector {
         if (playernameMatcher.matcher(playerName).matches()) {
             //only make a API call if the name is valid existing mojang account
 
-            if (System.currentTimeMillis() - lastRateLimit < 1_000 * 60 * 10) {
-                plugin.getLogger().fine("STILL WAITING FOR RATE_LIMIT - TRYING Third-party API");
+            if (requests.size() >= rateLimit || System.currentTimeMillis() - lastRateLimit < 1_000 * 60 * 10) {
+//                plugin.getLogger().fine("STILL WAITING FOR RATE_LIMIT - TRYING Third-party API");
                 return getUUIDFromAPI(playerName);
             }
+
+            requests.put(new Object(), new Object());
 
             try {
                 HttpsURLConnection connection = getConnection(UUID_LINK + playerName);
@@ -90,13 +102,13 @@ public abstract class MojangApiConnector {
                         return getUUIDFromJson(line);
                     }
                 } else if (connection.getResponseCode() == RATE_LIMIT_CODE) {
-                    plugin.getLogger().info("RATE_LIMIT REACHED - TRYING SECOND API");
+                    logger.info("RATE_LIMIT REACHED - TRYING THIRD-PARTY API");
                     lastRateLimit = System.currentTimeMillis();
                     return getUUIDFromAPI(playerName);
                 }
                 //204 - no content for not found
             } catch (Exception ex) {
-                plugin.getLogger().log(Level.SEVERE, "Failed to check if player has a paid account", ex);
+                logger.log(Level.SEVERE, "Failed to check if player has a paid account", ex);
             }
             //this connection doesn't need to be closed. So can make use of keep alive in java
         }
@@ -105,10 +117,6 @@ public abstract class MojangApiConnector {
     }
 
     public UUID getUUIDFromAPI(String playerName) {
-        if (!playernameMatcher.matcher(playerName).matches()) {
-            return null;
-        }
-
         try {
             HttpURLConnection httpConnection = (HttpURLConnection) new URL(MCAPI_UUID_URL + playerName).openConnection();
             httpConnection.addRequestProperty("Content-Type", "application/json");
@@ -122,10 +130,10 @@ public abstract class MojangApiConnector {
             BufferedReader reader = new BufferedReader(new InputStreamReader(httpConnection.getInputStream()));
             String line = reader.readLine();
             if (line != null && !line.equals("null")) {
-                return getUUIDFromJsonAPI(line);
+                return getUUIDFromJson(line);
             }
         } catch (IOException iOException) {
-            plugin.getLogger().log(Level.SEVERE, "Tried converting name->uuid from third-party api", iOException);
+            logger.log(Level.SEVERE, "Tried converting name->uuid from third-party api", iOException);
         }
 
         return null;
@@ -134,8 +142,6 @@ public abstract class MojangApiConnector {
     public abstract boolean hasJoinedServer(Object session, String serverId);
 
     protected abstract UUID getUUIDFromJson(String json);
-
-    protected abstract UUID getUUIDFromJsonAPI(String json);
 
     protected HttpsURLConnection getConnection(String url) throws IOException {
         HttpsURLConnection connection = (HttpsURLConnection) new URL(url).openConnection();