+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 * Added support for making requests to Mojang from different IPv4 addresses
 * Added us.mcapi.com as third-party APIs to workaround rate-limits
 * Fixed NPE in BungeeCord on cracked session
+* Fixed skin applies if premium uuid is deactivated
 
 ######1.6.2